Made the components routes proper
