+++ client/src/App.js
@@ -44,10 +44,10 @@
 
 import React from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { CreatePostForm } from "./components/CreatePostForm";
+import { CreatePostForm } from "./components/Home/CreatePostForm";
 import { Row, Col, Container } from "react-bootstrap";
-import Login from "./components/Login";
-import SignUp from "./components/SignUp";
+import Login from "./pages/Login";
+import SignUp from "./pages/SignUp";
 import { UserAuthContextProvider } from "./context/UserAuthContext";
 import ProtectedRoute from "./components/ProtectedRoute";
 
@@ -62,7 +62,7 @@ function App() {
                 <Route path="/login" element={<Login />} />
                 <Route path="/signup" element={<SignUp />} />
                 <Route
-                  path="/home"
+                  path="/home/*"
                   element={<ProtectedRoute val={1}/>}
                 ></Route>
                 <Route
@@ -73,10 +73,10 @@ function App() {
                   path="/profile/edit"
                   element={<ProtectedRoute val={3} />}
                 />
-                <Route
+                {/* <Route
                   path="/home/spost"
                   element={<ProtectedRoute val={4} />}
-                />
+                /> */}
                 {/* <Route path="/home/post" element={<CreatePostForm />} /> */}
               </Routes>
             </UserAuthContextProvider>
