+++ client/src/components/MajorComponents/LeftComponent.js
