+++ client/src/components/MajorComponents/MiddleComponent.js
