+++ client/src/components/MajorComponents/RightComponent.js
