+++ client/src/components/Profile/AddUserDetailForm.js
@@ -1,9 +1,5 @@
-import { useState, useEffect } from "react";
-import {
-  addUserDetails,
-  getUserDetails,
-  editUserDetails,
-} from "../service/userDetailsAPI";
+import { useEffect, useState } from "react";
+import { addUserDetails } from "../../service/userDetailsAPI";
 import {
   FormGroup,
   FormControl,
@@ -17,8 +13,9 @@ import {
   RadioGroup,
   FormLabel,
 } from "@mui/material";
-import { useUserAuth } from "../context/UserAuthContext";
-import { useNavigate } from "react-router-dom";
+import {  useNavigate } from "react-router-dom";
+import { useUserAuth } from "../../context/UserAuthContext";
+import { updateFlag } from "../../service/loginUserAPI";
 
 const Container = styled(FormGroup)`
   width: 50%;
@@ -40,82 +37,49 @@ const defaultValue = {
   Introduction: "",
 };
 
-export const EditUserDetailForm = () => {
+export const AddUserDetailForm = () => {
+  const PORT = 8000;
+
+  const { user: userCurr } = useUserAuth();
+
   const [user, setUser] = useState(defaultValue);
   const [value, setValue] = useState(0);
-  const { user: userCurr } = useUserAuth();
-  const [id, setId] = useState("yac984@gmail.byn");
+
   const navigate = useNavigate();
 
   const onValueChange = (e) => {
     if (e.target.name == "Job_Type") {
       setValue(e.target.value);
-      console.log(e.target.value);
     }
-    console.log(user);
     setUser({ ...user, [e.target.name]: e.target.value });
   };
 
-  useEffect(() => {
-    if (!userCurr) {
-      console.log("RETURN USER");
-      return;
-    }
-    if (userCurr.email == id) {
-      loadUserDetails();
-      return;
-    }
-    if (userCurr) {
-      setId(userCurr.email);
-    }
-  }, [userCurr, id]);
-
-  const loadUserDetails = async () => {
-    console.log("INSIDE LOAD   ", id);
-    const data = {
-      id: id,
-    };
-    const response = await getUserDetails(data);
-    setUser(response.data[0]);
-  };
   const handleSubmit = async () => {
-    user._id = id;
-    console.log("INSIDE HANDLE SUBMIT", user);
+    user._id = userCurr.email;
+    await addUserDetails(user);
     const data = {
-      id: id,
+      email: userCurr.email,
+      flag: true,
     };
-    await editUserDetails(user, data);
+    await updateFlag(data);
     navigate("/home");
   };
 
   return (
     <>
       <Container>
-        <div>{userCurr && userCurr.email}</div>
-        <Typography variant="h2">Edit Your Details</Typography>
+        <Typography variant="h2">Add Your Details</Typography>
         <FormControl>
           <InputLabel>Name</InputLabel>
-          <Input
-            onChange={(e) => onValueChange(e)}
-            name="Name"
-            value={user.Name}
-          />
+          <Input onChange={(e) => onValueChange(e)} name="Name" />
         </FormControl>
         <FormControl>
           <InputLabel>Age</InputLabel>
-          <Input
-            onChange={(e) => onValueChange(e)}
-            name="Age"
-            value={user.Age}
-          />
+          <Input onChange={(e) => onValueChange(e)} name="Age" />
         </FormControl>
         <FormControl>
           <InputLabel>UserName</InputLabel>
-          <Input
-            onChange={(e) => onValueChange(e)}
-            name="UserName"
-            value={user.UserName}
-          />
+          <Input onChange={(e) => onValueChange(e)} name="UserName" />
         </FormControl>
         <FormControl>
           <FormLabel id="demo-controlled-radio-buttons-group">
@@ -124,7 +88,7 @@ export const EditUserDetailForm = () => {
           <RadioGroup
             aria-labelledby="demo-controlled-radio-buttons-group"
             name="Job_Type"
-            value={user.Job_Type}
+            value={value}
             onChange={(e) => onValueChange(e)}
           >
             <FormControlLabel
@@ -138,35 +102,19 @@ export const EditUserDetailForm = () => {
         </FormControl>
         <FormControl>
           <InputLabel>City</InputLabel>
-          <Input
-            onChange={(e) => onValueChange(e)}
-            name="City"
-            value={user.City}
-          />
+          <Input onChange={(e) => onValueChange(e)} name="City" />
         </FormControl>
         <FormControl>
           <InputLabel>State</InputLabel>
-          <Input
-            onChange={(e) => onValueChange(e)}
-            name="State"
-            value={user.State}
-          />
+          <Input onChange={(e) => onValueChange(e)} name="State" />
         </FormControl>
         <FormControl>
           <InputLabel>Pincode</InputLabel>
-          <Input
-            onChange={(e) => onValueChange(e)}
-            name="Pincode"
-            value={user.Pincode}
-          />
+          <Input onChange={(e) => onValueChange(e)} name="Pincode" />
         </FormControl>
         <FormControl>
           <InputLabel>Introduction About Yourself</InputLabel>
-          <Input
-            onChange={(e) => onValueChange(e)}
-            name="Introduction"
-            value={user.Introduction}
-          />
+          <Input onChange={(e) => onValueChange(e)} name="Introduction" />
         </FormControl>
         <FormControl>
           <Button variant="contained" onClick={() => handleSubmit()}>
