+++ client/src/components/Profile/EditUserDetailForm.js
@@ -1,5 +1,9 @@
-import { useEffect, useState } from "react";
-import { addUserDetails } from "../service/userDetailsAPI";
+import { useState, useEffect } from "react";
+import {
+  addUserDetails,
+  getUserDetails,
+  editUserDetails,
+} from "../../service/userDetailsAPI";
 import {
   FormGroup,
   FormControl,
@@ -13,9 +17,8 @@ import {
   RadioGroup,
   FormLabel,
 } from "@mui/material";
-import {  useNavigate } from "react-router-dom";
-import { useUserAuth } from "../context/UserAuthContext";
-import { updateFlag } from "../service/loginUserAPI";
+import { useUserAuth } from "../../context/UserAuthContext";
+import { useNavigate } from "react-router-dom";
 
 const Container = styled(FormGroup)`
   width: 50%;
@@ -37,49 +40,82 @@ const defaultValue = {
   Introduction: "",
 };
 
-export const AddUserDetailForm = () => {
-  const PORT = 8000;
-
-  const { user: userCurr } = useUserAuth();
-
+export const EditUserDetailForm = () => {
   const [user, setUser] = useState(defaultValue);
   const [value, setValue] = useState(0);
-
+  const { user: userCurr } = useUserAuth();
+  const [id, setId] = useState("yac984@gmail.byn");
   const navigate = useNavigate();
 
   const onValueChange = (e) => {
     if (e.target.name == "Job_Type") {
       setValue(e.target.value);
+      console.log(e.target.value);
     }
+    console.log(user);
     setUser({ ...user, [e.target.name]: e.target.value });
   };
 
-  const handleSubmit = async () => {
-    user._id = userCurr.email;
-    await addUserDetails(user);
+  useEffect(() => {
+    if (!userCurr) {
+      console.log("RETURN USER");
+      return;
+    }
+    if (userCurr.email == id) {
+      loadUserDetails();
+      return;
+    }
+    if (userCurr) {
+      setId(userCurr.email);
+    }
+  }, [userCurr, id]);
+
+  const loadUserDetails = async () => {
+    console.log("INSIDE LOAD   ", id);
     const data = {
-      email: userCurr.email,
-      flag: true,
+      id: id,
     };
-    await updateFlag(data);
+    const response = await getUserDetails(data);
+    setUser(response.data[0]);
+  };
+  const handleSubmit = async () => {
+    user._id = id;
+    console.log("INSIDE HANDLE SUBMIT", user);
+    const data = {
+      id: id,
+    };
+    await editUserDetails(user, data);
     navigate("/home");
   };
 
   return (
     <>
       <Container>
-        <Typography variant="h2">Add Your Details</Typography>
+        <div>{userCurr && userCurr.email}</div>
+        <Typography variant="h2">Edit Your Details</Typography>
         <FormControl>
           <InputLabel>Name</InputLabel>
-          <Input onChange={(e) => onValueChange(e)} name="Name" />
+          <Input
+            onChange={(e) => onValueChange(e)}
+            name="Name"
+            value={user.Name}
+          />
         </FormControl>
         <FormControl>
           <InputLabel>Age</InputLabel>
-          <Input onChange={(e) => onValueChange(e)} name="Age" />
+          <Input
+            onChange={(e) => onValueChange(e)}
+            name="Age"
+            value={user.Age}
+          />
         </FormControl>
         <FormControl>
           <InputLabel>UserName</InputLabel>
-          <Input onChange={(e) => onValueChange(e)} name="UserName" />
+          <Input
+            onChange={(e) => onValueChange(e)}
+            name="UserName"
+            value={user.UserName}
+          />
         </FormControl>
         <FormControl>
           <FormLabel id="demo-controlled-radio-buttons-group">
@@ -88,7 +124,7 @@ export const AddUserDetailForm = () => {
           <RadioGroup
             aria-labelledby="demo-controlled-radio-buttons-group"
             name="Job_Type"
-            value={value}
+            value={user.Job_Type}
             onChange={(e) => onValueChange(e)}
           >
             <FormControlLabel
@@ -102,19 +138,35 @@ export const AddUserDetailForm = () => {
         </FormControl>
         <FormControl>
           <InputLabel>City</InputLabel>
-          <Input onChange={(e) => onValueChange(e)} name="City" />
+          <Input
+            onChange={(e) => onValueChange(e)}
+            name="City"
+            value={user.City}
+          />
         </FormControl>
         <FormControl>
           <InputLabel>State</InputLabel>
-          <Input onChange={(e) => onValueChange(e)} name="State" />
+          <Input
+            onChange={(e) => onValueChange(e)}
+            name="State"
+            value={user.State}
+          />
         </FormControl>
         <FormControl>
           <InputLabel>Pincode</InputLabel>
-          <Input onChange={(e) => onValueChange(e)} name="Pincode" />
+          <Input
+            onChange={(e) => onValueChange(e)}
+            name="Pincode"
+            value={user.Pincode}
+          />
         </FormControl>
         <FormControl>
           <InputLabel>Introduction About Yourself</InputLabel>
-          <Input onChange={(e) => onValueChange(e)} name="Introduction" />
+          <Input
+            onChange={(e) => onValueChange(e)}
+            name="Introduction"
+            value={user.Introduction}
+          />
         </FormControl>
         <FormControl>
           <Button variant="contained" onClick={() => handleSubmit()}>
