+++ client/src/config/firebase.js
